StorageManifestFactory returns an either
