+++ archive/registrar/src/main/scala/uk/ac/wellcome/platform/archive/registrar/factories/StorageManifestFactory.scala
@@ -1,27 +1,24 @@
 package uk.ac.wellcome.platform.archive.registrar.factories
-import cats.implicits._
+import akka.stream.Materializer
+import akka.stream.scaladsl.{Sink, StreamConverters}
 import com.amazonaws.services.s3.AmazonS3
 import grizzled.slf4j.Logging
 import uk.ac.wellcome.platform.archive.common.models.BagLocation
-import uk.ac.wellcome.platform.archive.registrar.models
+import uk.ac.wellcome.platform.archive.registrar.flows.FileSplitterFlow
 import uk.ac.wellcome.platform.archive.registrar.models._
-import uk.ac.wellcome.platform.archive.registrar.models.errors.{
-  FileNotFoundError,
-  InvalidBagManifestError,
-  RegistrarError
-}
 import uk.ac.wellcome.storage.ObjectLocation
 
-import scala.util.Try
+import scala.concurrent.duration._
+import scala.concurrent.{Await, ExecutionContext}
 
 object StorageManifestFactory extends Logging {
-  def create(bagLocation: BagLocation)(
-    implicit s3Client: AmazonS3): Either[RegistrarError, StorageManifest] = {
+  def create(bagLocation: BagLocation)(implicit s3Client: AmazonS3,
+                                       materializer: Materializer,
+                                       executionContext: ExecutionContext): Either[Throwable, StorageManifest] = {
 
     val algorithm = "sha256"
 
-    val manifestTupleFuture =
-      getBagItems(bagLocation, s"manifest-$algorithm.txt", " +")
+    val manifestTupleFuture = getTuples(bagLocation, s"manifest-$algorithm.txt", " +")
 //    val tagManifestTupleFuture = getTuples(s"tagmanifest-$algorithm.txt", " +")
 
     val sourceIdentifier = SourceIdentifier(
@@ -41,11 +38,11 @@ object StorageManifestFactory extends Logging {
       )
     )
 
-    for {
+    val eventualManifest = for {
       manifestTuples <- manifestTupleFuture
       fileManifest = FileManifest(
         ChecksumAlgorithm(algorithm),
-        manifestTuples
+        createBagDigestFiles(manifestTuples).toList
       )
       //tagManifestTuples <- tagManifestTupleFuture
       tagManifest = TagManifest(
@@ -61,35 +58,25 @@ object StorageManifestFactory extends Logging {
         tagManifest = tagManifest,
         locations = List(location)
       )
+    Right(Await.result(eventualManifest,10 seconds))
   }
 
-  def getBagItems(bagLocation: BagLocation, name: String, delimiter: String)(
-    implicit s3Client: AmazonS3)
-    : Either[RegistrarError, List[BagDigestFile]] = {
-    val location = createBagItMetaFileLocation(bagLocation, name)
-    val triedLines = Try(s3Client.getObject(location.namespace, location.key))
-      .map(_.getObjectContent)
-      .toEither
-      .leftMap(_ => FileNotFoundError(bagLocation))
-      .map(
-        inputStream =>
-          scala.io.Source
-            .fromInputStream(inputStream)
-            .mkString
-            .split("\n")
-            .filter(_.nonEmpty)
-            .toList)
+  def getTuples(bagLocation: BagLocation, name: String, delimiter: String)(implicit s3Client: AmazonS3, materializer: Materializer) = {
+    val location = createBagItMetaFileLocation(bagLocation,name)
+    s3LocationToSource(location)
+      .via(FileSplitterFlow(delimiter))
+      .runWith(Sink.seq)
 
-    triedLines.flatMap { lines: List[String] =>
-      val value: Either[InvalidBagManifestError, List[BagDigestFile]] =
-        lines.map { line =>
-          parseManifestLine(line, delimiter, bagLocation, name)
-        }.sequence
-      value
+  }
+
+  def createBagDigestFiles(digestLines: Seq[(String, String)]) = {
+    digestLines.map {
+      case (checksum, path) =>
+        BagDigestFile(Checksum(checksum), BagFilePath(path))
     }
   }
 
-  def createBagItMetaFileLocation(bagLocation: BagLocation, name: String) =
+  def createBagItMetaFileLocation(bagLocation: BagLocation,name: String) =
     ObjectLocation(
       bagLocation.storageNamespace,
       List(
@@ -99,22 +86,9 @@ object StorageManifestFactory extends Logging {
       ).mkString("/")
     )
 
-  def parseManifestLine(
-    fileChunk: String,
-    delimiter: String,
-    job: BagLocation,
-    manifestName: String): Either[InvalidBagManifestError, BagDigestFile] = {
-    val splitChunk = fileChunk.split(delimiter).map(_.trim)
-
-    splitChunk match {
-      case Array(checksum: String, key: String) =>
-        Right(
-          models.BagDigestFile(
-            Checksum(checksum),
-            BagFilePath(key)
-          ))
-      case _ =>
-        Left(InvalidBagManifestError(job, manifestName))
-    }
+  def s3LocationToSource(location: ObjectLocation)(implicit s3Client: AmazonS3) = {
+    debug(s"Attempting to get ${location.namespace}/${location.key}")
+    val s3Object = s3Client.getObject(location.namespace, location.key)
+    StreamConverters.fromInputStream(() => s3Object.getObjectContent)
   }
 }
