+++ archive/registrar/src/main/scala/uk/ac/wellcome/platform/archive/registrar/models/errors/RegistrarError.scala
@@ -3,8 +3,4 @@ import uk.ac.wellcome.platform.archive.common.models.BagLocation
 
 sealed trait RegistrarError
 
-case class FileNotFoundError(bagLocation: BagLocation) extends RegistrarError
-
-case class InvalidBagManifestError(bagLocation: BagLocation,
-                                   manifestName: String)
-    extends RegistrarError
+case class FileNotFoundError(bagLocation: BagLocation)
