+++ archive/registrar/src/test/scala/uk/ac/wellcome/platform/archive/registrar/RegistrarFeatureTest.scala
@@ -49,8 +49,6 @@ class RegistrarFeatureTest
           storageBucket) { bagLocation =>
           registrar.run()
 
-          implicit val _ = s3Client
-
             eventually {
               val messages = listMessagesReceivedFromSNS(topic)
               messages should have size 1
