+++ archive/registrar/src/test/scala/uk/ac/wellcome/platform/archive/registrar/fixtures/BagLocationFixtures.scala
@@ -1,10 +1,6 @@
 package uk.ac.wellcome.platform.archive.registrar.fixtures
 import uk.ac.wellcome.platform.archive.common.fixtures.{BagIt, FileEntry}
-import uk.ac.wellcome.platform.archive.common.models.{
-  BagLocation,
-  BagPath,
-  DigitisedStorageType
-}
+import uk.ac.wellcome.platform.archive.common.models.{BagLocation, BagPath, DigitisedStorageType}
 import uk.ac.wellcome.storage.fixtures.S3
 import uk.ac.wellcome.storage.fixtures.S3.Bucket
 import uk.ac.wellcome.test.fixtures.TestWith
